pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pneuma"
version = "1.0.0"
description = "Zero-Friction, Auto-Curated AI Memory Bridge"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "mempalace>=2.0.0",
    "slack-sdk>=3.41.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Symbol-aware code mining via tree-sitter. Without this extra, the miner
# falls back to character-based chunking automatically.
mining = [
    "tree-sitter>=0.21",
    "tree-sitter-language-pack>=0.2",
]

[project.scripts]
pneuma = "core.cli:main"
pneuma-mcp = "mcp_server.server:main"

[tool.setuptools.packages.find]
include = ["core*", "mcp_server*", "chat_bot*"]
